+++ src/threads.js
@@ -61,7 +61,7 @@ StageMorph, SpriteMorph, StagePrompterMorph, Note, modules, isString, copy, Map,
 isNil, WatcherMorph, List, ListWatcherMorph, alert, console, TableMorph, BLACK,
 TableFrameMorph, ColorSlotMorph, isSnapObject, newCanvas, Symbol, SVG_Costume*/
 
-modules.threads = '2020-December-19';
+modules.threads = '2020-December-22';
 
 var ThreadManager;
 var Process;
@@ -5855,9 +5855,9 @@ Process.prototype.reportNewCostume = function (pixels, width, height, name) {
     for (i = 0; i < src.length; i += 1) {
         px = src[i].itemsArray();
         for (k = 0; k < 3; k += 1) {
-            dta.data[(i * 4) + k] = px[k];
+            dta.data[(i * 4) + k] = +px[k];
         }
-        dta.data[i * 4 + 3] = (px[3] === undefined ? 255 : px[3]);
+        dta.data[i * 4 + 3] = (px[3] === undefined ? 255 : +px[3]);
     }
     ctx.putImageData(dta, 0, 0);
     return new Costume(